Modify package br.com.monkey.ecx to br.com.monkey.sdk
