--- src/main/java/br/com/monkey/ecx/annotation/SearchParameter.java
@@ -1,14 +1,0 @@
-package br.com.monkey.ecx.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SearchParameter {
-
-	String value() default "search";
-
-}
