--- src/main/java/br/com/monkey/ecx/configuration/SearchBySpecificationAutoConfiguration.java
@@ -1,17 +1,0 @@
-package br.com.monkey.ecx.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-public class SearchBySpecificationAutoConfiguration implements WebMvcConfigurer {
-
-	@Override
-	public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-		resolvers.add(new SearchParameterArgumentResolver());
-	}
-
-}
