+++ src/main/java/br/com/monkey/sdk/annotation/SearchParameter.java
@@ -1,0 +1,14 @@
+package br.com.monkey.sdk.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SearchParameter {
+
+	String value() default "search";
+
+}
