+++ src/main/java/br/com/monkey/sdk/core/exception/MonkeyRuntimeException.java
@@ -1,0 +1,55 @@
+package br.com.monkey.sdk.core.exception;
+
+import br.com.monkey.sdk.core.model.MessageType;
+
+import java.util.List;
+
+public class MonkeyRuntimeException extends RuntimeException {
+
+	private static final long serialVersionUID = 1L;
+
+	private List<String> notifications;
+
+	private MessageType messageType;
+
+	private Integer errorCode;
+
+	public MonkeyRuntimeException() {
+	}
+
+	public MonkeyRuntimeException(MessageType messageType, List<String> notifications, Integer errorCode) {
+		super(new Throwable(notifications.toString()));
+		this.messageType = messageType;
+		this.notifications = notifications;
+		this.errorCode = errorCode;
+	}
+
+	public MonkeyRuntimeException(MessageType messageType, List<String> notifications, Integer errorCode,
+			boolean enableSuppression) {
+		super(notifications.toString(), null, enableSuppression, false);
+		this.messageType = messageType;
+		this.notifications = notifications;
+		this.errorCode = errorCode;
+	}
+
+	public MonkeyRuntimeException(MessageType messageType, List<String> notifications, Throwable cause,
+			Integer errorCode) {
+		super(cause);
+		this.messageType = messageType;
+		this.notifications = notifications;
+		this.errorCode = errorCode;
+	}
+
+	public MessageType getMessageType() {
+		return messageType;
+	}
+
+	public List<String> getNotifications() {
+		return notifications;
+	}
+
+	public int getHttpErrorCode() {
+		return errorCode;
+	}
+
+}
