+++ src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=br.com.monkey.ecx.configuration.SearchBySpecificationAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=br.com.monkey.sdk.configuration.SearchBySpecificationAutoConfiguration
