--- src/test/java/br/com/monkey/ecx/Category.java
@@ -1,29 +1,0 @@
-package br.com.monkey.ecx;
-
-import lombok.*;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import java.time.Instant;
-
-import static lombok.AccessLevel.PRIVATE;
-
-@Entity
-@Getter
-@Builder
-@ToString
-@EqualsAndHashCode
-@AllArgsConstructor
-@NoArgsConstructor(access = PRIVATE)
-class Category {
-
-	@Id
-	private Integer id;
-
-	private String name;
-
-	private Instant updatedAt;
-
-	private CategoryType type;
-
-}
