--- src/test/java/br/com/monkey/ecx/CategoryType.java
@@ -1,9 +1,0 @@
-package br.com.monkey.ecx;
-
-enum CategoryType {
-
-	TYPE,
-
-	TYPE2
-
-}
