--- src/test/java/br/com/monkey/ecx/MysqlIntegrationContainerConfiguration.java
@@ -1,23 +1,0 @@
-package br.com.monkey.ecx;
-
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.MySQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-@Testcontainers
-public class MysqlIntegrationContainerConfiguration {
-
-	@Container
-	public static final MySQLContainer<?> database = new MySQLContainer<>("mysql:8.0.32").withUsername("root")
-		.withReuse(true);
-
-	@DynamicPropertySource
-	static void databaseProperties(DynamicPropertyRegistry registry) {
-		registry.add("spring.datasource.url", database::getJdbcUrl);
-		registry.add("spring.datasource.username", database::getUsername);
-		registry.add("spring.datasource.password", database::getPassword);
-	}
-
-}
