--- src/test/java/br/com/monkey/ecx/ProductRepository.java
@@ -1,8 +1,0 @@
-package br.com.monkey.ecx;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
-interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
-
-}
