--- src/test/java/br/com/monkey/ecx/SpringSpecificationSearchApplication.java
@@ -1,13 +1,0 @@
-package br.com.monkey.ecx;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-class SpringSpecificationSearchApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringSpecificationSearchApplication.class, args);
-	}
-
-}
