+++ src/test/java/br/com/monkey/sdk/CategoryRepository.java
@@ -1,0 +1,8 @@
+package br.com.monkey.sdk;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+interface CategoryRepository extends JpaRepository<Category, Integer>, JpaSpecificationExecutor<Category> {
+
+}
