+++ src/test/java/br/com/monkey/sdk/CategoryType.java
@@ -1,0 +1,9 @@
+package br.com.monkey.sdk;
+
+enum CategoryType {
+
+	TYPE,
+
+	TYPE2
+
+}
