+++ src/test/java/br/com/monkey/sdk/ProductRepository.java
@@ -1,0 +1,8 @@
+package br.com.monkey.sdk;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
+
+}
