+++ src/test/java/br/com/monkey/sdk/SpringSpecificationSearchApplication.java
@@ -1,0 +1,13 @@
+package br.com.monkey.sdk;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+class SpringSpecificationSearchApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringSpecificationSearchApplication.class, args);
+	}
+
+}
