+++ src/test/resources/logback-test.xml
@@ -3,7 +3,7 @@
     <include resource="org/springframework/boot/logging/logback/defaults.xml"/>
     <include resource="org/springframework/boot/logging/logback/console-appender.xml"/>
 
-    <logger name="elk=br.com.monkey.ecx" level="INFO" additivity="true">
+    <logger name="elk=br.com.monkey.sdk" level="INFO" additivity="true">
         <appender-ref ref="CONSOLE"/>
     </logger>
